COMMUNICATION-547 - Inform the HUB of a CoT
Draft

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@50702 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/dlms/src/main/java/com/energyict/dlms/axrdencoding/util/DateTime.java
@@ -165,7 +165,7 @@ public class DateTime extends AbstractDataType {
 
         return
             new byte [] {
-                (byte) DLMSCOSEMGlobals.TYPEDESC_OCTET_STRING,
+                DLMSCOSEMGlobals.TYPEDESC_OCTET_STRING,
                 (byte) SIZE,
                 (byte) ((year & INT_HIGH_MASK ) >> BYTE_SIZE),
                 (byte) (year & INT_LOW_MASK),
